+++ services/cd-service/pkg/service/overview.go
@@ -19,7 +19,9 @@ package service
 import (
 	"context"
 	"errors"
+	"fmt"
 	"os"
+	"path/filepath"
 	"sync"
 	"sync/atomic"
 
@@ -139,6 +141,15 @@ func (o *OverviewServiceServer) getOverview(
 								Message: lock.Message,
 								Commit:  lockCommit,
 								LockId:  lockId,
+							}
+						}
+					}
+					if config.ArgoCd != nil {
+						if syncWindows, err := transformSyncWindows(config.ArgoCd.SyncWindows, appName); err != nil {
+							return nil, err
+						} else {
+							app.ArgoCD = &api.Environment_Application_ArgoCD{
+								SyncWindows: syncWindows,
 							}
 						}
 					}
@@ -271,4 +282,22 @@ func transformCommit(commit *git.Commit) *api.Commit {
 		AuthorEmail: author.Email,
 		AuthorTime:  timestamppb.New(author.When),
 	}
+}
+
+func transformSyncWindows(syncWindows []config.ArgoCdSyncWindow, appName string) ([]*api.Environment_Application_ArgoCD_SyncWindow, error) {
+	var envAppSyncWindows []*api.Environment_Application_ArgoCD_SyncWindow
+	for _, syncWindow := range syncWindows {
+		for _, pattern := range syncWindow.Apps {
+			if match, err := filepath.Match(pattern, appName); err != nil {
+				return nil, fmt.Errorf("failed to match app pattern %s of sync window to %s at %s with duration %s: %w", pattern, syncWindow.Kind, syncWindow.Schedule, syncWindow.Duration, err)
+			} else if match {
+				envAppSyncWindows = append(envAppSyncWindows, &api.Environment_Application_ArgoCD_SyncWindow{
+					Kind:     syncWindow.Kind,
+					Schedule: syncWindow.Schedule,
+					Duration: syncWindow.Duration,
+				})
+			}
+		}
+	}
+	return envAppSyncWindows, nil
 }