+++ services/frontend-service/src/ui/ConfirmationDialog.tsx
@@ -14,7 +14,7 @@ You should have received a copy of the GNU General Public License
 along with kuberpult.  If not, see <http://www.gnu.org/licenses/>.
 
 Copyright 2021 freiheit.com*/
-import { BatchAction, Lock } from '../api/api';
+import { BatchAction, Environment_Application_ArgoCD_SyncWindow, Lock } from '../api/api';
 import * as React from 'react';
 import { Button, Dialog, DialogTitle, IconButton, Typography, Snackbar, Alert, AlertTitle } from '@material-ui/core';
 import { useCallback, useContext } from 'react';
@@ -28,6 +28,7 @@ import {
     MoveToInboxRounded,
 } from '@material-ui/icons';
 import { ActionsCartContext } from './App';
+import { SyncWindow } from './ReleaseDialog';
 
 enum ActionTypes {
     Deploy,
@@ -95,10 +96,11 @@ export interface ConfirmationDialogProviderProps {
     locks?: [string, Lock][];
     undeployedUpstream?: string;
     fin?: () => void;
+    syncWindows?: Environment_Application_ArgoCD_SyncWindow[];
 }
 
 export const ConfirmationDialogProvider = (props: ConfirmationDialogProviderProps) => {
-    const { action, locks, fin, undeployedUpstream } = props;
+    const { action, locks, fin, undeployedUpstream, syncWindows } = props;
     const [openNotify, setOpenNotify] = React.useState(false);
     const [dialogOpen, setDialogOpen] = React.useState(false);
     const { actions, setActions } = useContext(ActionsCartContext);
@@ -182,6 +184,21 @@ export const ConfirmationDialogProvider = (props: ConfirmationDialogProviderProp
             <strong>Possible conflict with actions already in cart!</strong>
         </Alert>
     );
+    const syncWindowsMessage =
+        syncWindows && syncWindows.length > 0 ? (
+            <Alert variant="outlined" sx={{ m: 1 }} severity="warning">
+                <AlertTitle>ArgoCD sync windows are active for this application!</AlertTitle>
+                <p>Warning: This can delay deployment.</p>
+                <h3>Sync windows:</h3>
+                <ul>
+                    {syncWindows?.map((w, n) => (
+                        <li key={`${n}:${w}`}>
+                            <SyncWindow w={w} />
+                        </li>
+                    ))}
+                </ul>
+            </Alert>
+        ) : null;
 
     return (
         <>
@@ -211,6 +228,7 @@ export const ConfirmationDialogProvider = (props: ConfirmationDialogProviderProp
                 {deployLocks}
                 {undeployedUpstreamMessage}
                 {conflictMessage}
+                {syncWindowsMessage}
                 <span style={{ alignSelf: 'end' }}>
                     <Button onClick={closeDialog}>Cancel</Button>
                     <Button onClick={addAction}>Add anyway</Button>
@@ -252,7 +270,7 @@ export const GetActionDetails = (action: BatchAction): ActionDetails => {
             return {
                 type: ActionTypes.Deploy,
                 name: 'Deploy',
-                dialogTitle: 'Are you sure you want to deploy this version?',
+                dialogTitle: 'Please be aware:',
                 notMessageSuccess:
                     'Version ' +
                     action.action?.deploy.version +